fastapi-bq-accesslog/pyproject.toml
[project]
name = "fastapi-bq-accesslog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.115.0",
    "fastavro>=1.9.7",
    "gcloud-aio-pubsub>=6.0.1",
]

[tool.uv]
dev-dependencies = [
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest>=8.3.3",
]
